pyproject.toml
[project]
name = "kt-ds"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "bs4>=0.0.2",
    "faiss-cpu>=1.11.0",
    "kiwipiepy>=0.21.0",
    "langchain>=0.3.25",
    "langchain-chroma>=0.2.4",
    "langchain-cli>=0.0.36",
    "langchain-community>=0.3.24",
    "langchain-experimental>=0.3.4",
    "langchain-huggingface>=0.2.0",
    "langchain-mcp-adapters>=0.1.1",
    "langchain-ollama>=0.3.3",
    "langchain-openai>=0.3.18",
    "langchain-text-splitters>=0.3.8",
    "langgraph>=0.4.7",
    "openpyxl>=3.1.5",
    "pillow>=11.2.1",
    "pypdf>=5.5.0",
    "python-dotenv>=1.1.0",
    "ragas>=0.2.15",
    "rank-bm25>=0.2.2",
    "rapidfuzz>=3.13.0",
]
